move share price to coop config.py
