+++ tapir/coop/config.py
@@ -1,0 +1,4 @@
+from decimal import Decimal
+
+COOP_SHARE_PRICE = Decimal(100)
+COOP_ENTRY_AMOUNT = Decimal(10)
