+++ tapir/coop/views/statistics.py
@@ -7,13 +7,13 @@ from django.views import generic
 from django.views.generic import TemplateView
 
 from tapir.accounts.models import TapirUser
+from tapir.coop.config import COOP_SHARE_PRICE
 from tapir.coop.models import (
     ShareOwner,
     MemberStatus,
     DraftUser,
     ShareOwnership,
 )
-from tapir.settings import COOP_SHARE_PRICE
 
 
 class StatisticsView(PermissionRequiredMixin, generic.TemplateView):
